+++ lib/database/suggestproducts.ts
@@ -3,6 +3,7 @@
 import { revalidatePath } from "next/cache";
 import prisma from "./conn";
 import { requireUser } from "../requireUser";
+import { requireAdmin } from "../requireAdmin";
 
 export interface suggestProducts {
   id: string;
@@ -20,7 +21,6 @@ export interface suggestProducts {
 
 export async function getAllSuggestProducts() {
   try {
-
     const suggests = await prisma.suggestProducts.findMany({
       include: {
         product: {
@@ -55,7 +55,7 @@ export async function getAllSuggestProducts() {
 
 export async function addSuggestProducts(id: string) {
   try {
-    await requireUser()
+    await requireAdmin()
     await prisma.suggestProducts.create({
       data: {
         productId: id,
@@ -72,7 +72,7 @@ export async function addSuggestProducts(id: string) {
 
 export async function DeleteSuggestProduct(id: string) {
   try {
-    await requireUser()
+    await requireAdmin()
     await prisma.suggestProducts.delete({
       where: {
         id: id
